factor server build stage out to startup
