+++ tests/api/health_check.rs
@@ -1,0 +1,17 @@
+#[tokio::test]
+async fn health_check_works() {
+    // Arrange
+    let app = TestApp::spawn().await;
+    let client = reqwest::Client::new();
+
+    // Act
+    let response = client
+        .get(&format!("{}/health_check", &app.address))
+        .send()
+        .await
+        .expect("Failed to execute request");
+
+    // Assert
+    assert!(response.status().is_success());
+    assert_eq!(response.content_length(), Some(0));
+}
