+++ tests/api/subscriptions.rs
@@ -1,25 +1,3 @@
-mod helpers;
-
-use helpers::TestApp;
-
-#[tokio::test]
-async fn health_check_works() {
-    // Arrange
-    let app = TestApp::spawn().await;
-    let client = reqwest::Client::new();
-
-    // Act
-    let response = client
-        .get(&format!("{}/health_check", &app.address))
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    // Assert
-    assert!(response.status().is_success());
-    assert_eq!(response.content_length(), Some(0));
-}
-
 #[tokio::test]
 async fn subscribe_returns_a_200_for_valid_form_data() {
     // Arrange
